+++ src/components/islets/anime-watch-islets/anime-player.tsx
@@ -1,25 +1,13 @@
 "use client";
 import React, { useEffect } from "react";
-import axios from "axios";
 import { useSearchParams } from "next/navigation";
 import { useQuery, useQueryClient } from "@tanstack/react-query";
-import Player from "@/components/ui/player";
-import { API_HOST_CLIENT, GOGOANIME_ENDPOINT, ANIME } from "@/config";
-import { Spinner } from "@/components/ui/spinner";
-import Ripple from "@/components/ui/ripple";
-import { AnimeVideoDetailsType } from "@/lib/types";
+import { fetchAnimeStreamingLinks } from "@/lib/anime";
+import RenderVideoAnime from "./anime-render-video";
 
 interface AnimePlayerProps {
   episodeId: string;
   cover: string;
-}
-
-interface RenderVideoProps {
-  data: any;
-  cover: string;
-  isFetching: boolean;
-  status: string;
-  error: any;
 }
 
 const AnimePlayer: React.FC<AnimePlayerProps> = ({ episodeId, cover }) => {
@@ -34,18 +22,6 @@ const AnimePlayer: React.FC<AnimePlayerProps> = ({ episodeId, cover }) => {
     queryFn: () => fetchAnimeStreamingLinks(selectedEpisode),
     refetchOnWindowFocus: false,
   });
-
-  const fetchAnimeStreamingLinks = async (
-    episode: string
-  ): Promise<AnimeVideoDetailsType> => {
-    const { data } = await axios.get(
-      `${
-        API_HOST_CLIENT + ANIME + GOGOANIME_ENDPOINT
-      }/watch/${encodeURIComponent(episode)}`
-    );
-
-    return data;
-  };
 
   useEffect(() => {
     queryClient.invalidateQueries({
@@ -62,7 +38,7 @@ const AnimePlayer: React.FC<AnimePlayerProps> = ({ episodeId, cover }) => {
       />
 
       <div className="relative px-0 md:px-10 lg:px-16">
-        <RenderVideo
+        <RenderVideoAnime
           data={data}
           cover={cover}
           isFetching={isFetching}
@@ -73,49 +49,5 @@ const AnimePlayer: React.FC<AnimePlayerProps> = ({ episodeId, cover }) => {
     </div>
   );
 };
-
-function RenderVideo({
-  data,
-  cover,
-  isFetching,
-  error,
-  status,
-}: RenderVideoProps) {
-  if (status === "error") {
-    console.log(error);
-    return (
-      <div className="flex justify-center items-center h-[250px] sm:h-[330px] md:h-[450px] lg:h-[570px] xl:h-[660px]">
-        <div className="flex flex-col items-center space-y-2">
-          <img src="/images/error.png" alt="Error" className="w-24 h-24" />
-          <h1>There&apos;s an error playing video.</h1>
-        </div>
-      </div>
-    );
-  }
-
-  if (isFetching) {
-    return (
-      <div className="flex justify-center items-center h-[260px] sm:h-[340px] md:h-[460px] lg:h-[580px] xl:h-[670px]">
-        {/* <Spinner className="text-orange-400" size={40} /> */}
-        <Ripple />
-      </div>
-    );
-  }
-
-  if (status === "success") {
-    return (
-      <div className="w-full 2xl:w-3/4 m-auto mt-0">
-        <Player
-          src={
-            data?.sources?.filter((item: any) => item.quality === "default")[0]
-              .url
-          }
-        />
-      </div>
-    );
-  }
-
-  return null;
-}
 
 export default AnimePlayer;
